migrated confirmby tests into separate files
