+++ common/utils/utils.test.ts
@@ -1,7 +1,6 @@
 import { describe, expect, test } from '@jest/globals';
-import { useConfirmByState } from '../components/hooks/useConfirmByState';
-import { ConfirmByState } from '../common/types';
-import { getConfirmByState } from '../common/utils';
+import { ConfirmByState } from '../types';
+import { getConfirmByState } from './utils';
 
 const mockSystemDate = (d: Date) => {
   jest.useFakeTimers().setSystemTime(d.getTime());
@@ -13,13 +12,11 @@ describe('confirm by tests', () => {
   test('before', () => {
     // if the current date is jan 1, and the confirm date is jan 21, it is Before confirm date
     mockSystemDate(new Date('1/1/2022'));
-    expect(useConfirmByState({ confirmBy: confirmDate })).toBe(ConfirmByState.Before);
     expect(getConfirmByState(confirmDate)).toBe(ConfirmByState.Before);
   });
 
   test('after', () => {
     mockSystemDate(new Date('2/1/2022'));
-    expect(useConfirmByState({ confirmBy: confirmDate })).toBe(ConfirmByState.After);
     expect(getConfirmByState(confirmDate)).toBe(ConfirmByState.After);
   });
 });
