+++ components/application/ApplicationForm.tsx
@@ -11,7 +11,7 @@ import { Questions, Sections } from '../../common/questions';
 import { Alert, Button, Form, notification } from 'antd';
 import useSWR from 'swr';
 import { format } from '../dashboard/status-dialogue/StatusDialogue';
-import { isAfterRegistrationClosed, isBeforeRegistrationOpens } from '../../common/utils';
+import { isAfterRegistrationClosed, isBeforeRegistrationOpens } from '../../common/utils/utils';
 import { useWarnIfUnsavedChanges } from '../hooks/useWarnIfUnsavedChanges';
 import { FormSectionsAndQuestions } from './FormSectionsAndQuestions';
