+++ components/dashboard/DecisionStatusDialogue.tsx
@@ -1,6 +1,6 @@
 import { ConfirmByState, DecisionStatus } from '../../common/types';
 import React from 'react';
-import { assertUnreachable } from '../../common/utils';
+import { assertUnreachable } from '../../common/utils/utils';
 import { Alert, Button } from 'antd';
 import { format } from './status-dialogue/StatusDialogue';
 import Link from 'next/link';
