+++ pages/api/v1/post-acceptance.ts
@@ -5,7 +5,7 @@ import { makeQuestionResponseSchemas } from '../../../server/validators';
 import Joi from 'joi';
 import { connectToDatabase } from '../../../server/mongoDB';
 import { assumeLoggedInGetEmail, protect } from '../../../server/protect';
-import { getConfirmByState } from '../../../common/utils';
+import { getConfirmByState } from '../../../common/utils/utils';
 import { PostAcceptanceFormQuestions } from '../../../common/questions';
 
 const postAcceptanceHandler: NextApiHandler = async (req, res) => {
